% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GOTnames}
\alias{GOTnames}
\title{Game of Thrones character names + word lengths dats}
\format{A matrix with 100 rows and 3 columns}
\usage{
GOTnames
}
\value{
A matrix with 100 rows and 3 columns
}
\description{
Game of Thrones character names + word lengths dats
}
\keyword{datasets}

